Share SSH session for mirroring Motivation: When mirroring, SSH session is currently  established a couple of times. Establishing an SSH session is a highly resource-intensive task. If would be nice if we can reduce it.
Modifications:

Result:
