+++ server-mirror-git/src/main/java/com/linecorp/centraldogma/server/internal/mirror/GitMirrorProvider.java
@@ -43,10 +43,16 @@ public final class GitMirrorProvider implements MirrorProvider {
         }
 
         switch (scheme) {
-            case SCHEME_GIT:
-            case SCHEME_GIT_SSH:
+            case SCHEME_GIT_SSH: {
+                final String[] components = split(remoteUri, "git");
+                return new SshGitMirror(context.schedule(), context.direction(), context.credential(),
+                                        context.localRepo(), context.localPath(),
+                                        URI.create(components[0]), components[1], components[2],
+                                        context.gitignore());
+            }
             case SCHEME_GIT_HTTP:
             case SCHEME_GIT_HTTPS:
+            case SCHEME_GIT:
             case SCHEME_GIT_FILE: {
                 final String[] components = split(remoteUri, "git");
                 return new GitMirror(context.schedule(), context.direction(), context.credential(),
